+++ docs/release_notes.rst
@@ -8,10 +8,12 @@ Release Notes
 1.11.1 (Unreleased)
 ---------------------------------------------------------------------------------------
 
-Bug fixes
-^^^^^^^^^
+Performance Improvements
+^^^^^^^^^^^^^^^^^^^^^^^^
 
-* You know there will be some...
+* Parquet outputs are now written with ZSTD compression at level 3, instead of
+  DuckDB's default Snappy compression, for smaller output files. See issue
+  :issue:`461` and PR :pr:`462`.
 
 .. _release-v1-11-0: